@@ -47,7 +47,7 @@
     </header>
     <div class="container">
         <partial name="_CookieConsentPartial" />
-        <script src="~/lib/signalr/dist/browser/signalr.js"></script>
+        <script src="~/lib/microsoft/signalr/dist/browser/signalr.js"></script>
         <script src="~/lib/signalr.eventaggregatorproxy/signalR.eventAggregator.js" asp-append-version="true"></script>
         <script src="~/lib/knockout/build/output/knockout-latest.js" asp-append-version="true"></script>  
         <script src="/eventAggregation/events" asp-append-version="true"></script>
