@@ -1,11 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Text.Json;
 
 namespace SignalR.EventAggregatorProxy.Model
 {
     public class Subscription
     {
-        public Subscription(Type eventType, string connectionId, string username, dynamic constraint, int? constraintId, IList<Type> genericArguments)
+        public Subscription(Type eventType, string connectionId, string username, JsonElement constraint, int? constraintId, IList<Type> genericArguments)
         {
             EventType = eventType;
             ConnectionId = connectionId;
@@ -18,7 +19,7 @@ namespace SignalR.EventAggregatorProxy.Model
         public Type EventType { get; set; }
         public string ConnectionId { get; set; }
         public string Username { get; set; }
-        public dynamic Constraint { get; set; }
+        public JsonElement Constraint { get; set; }
         public IList<Type> GenericArguments { get; set; }
         public int? ConstraintId { get; set; }
     }
